Optimize BEEFY relay
